refactor: Wave 2 (foundation) - SFNetworkHelper to Core/Net/
Completes the shared networking foundation alongside SFRCO (Wave 1): RCO + the
authority/network helper now both live under Core/Net/. Both include forms
rewritten (qualified Core/Helpers/ + one bare); Shipping module links clean.

Remaining Wave 2: extract the construct/spec hooks from SFGameInstanceModule into
Core/Net/ (needs in-game MP validation - hook registration, compile != correct),
the [MP-*] re-tag, and per-feature divergence maps.
